requirements.txt
numpy~=1.26.4
pandas~=2.2.1
plotly~=5.20.0
tqdm~=4.66.2
scipy~=1.12.0
scikit-learn~=1.4.1.post1
gensim~=4.3.2
umap-learn~=0.5.6